# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineageEdge": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "edge_type": {
            "enum": [
              "derived_from",
              "converted_to",
              "candidate_for",
              "mapped_to",
              "reviewed_by",
              "influenced_by",
              "validated_against",
              "rendered_as",
              "contained_in",
              "verified_by"
            ],
            "title": "Edge Type",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "source_node_id": {
            "title": "Source Node Id",
            "type": "string"
          },
          "target_node_id": {
            "title": "Target Node Id",
            "type": "string"
          }
        },
        "required": [
          "edge_id",
          "source_node_id",
          "target_node_id",
          "edge_type"
        ],
        "title": "LineageEdge",
        "type": "object"
      },
      "LineageEvidence": {
        "additionalProperties": false,
        "properties": {
          "artifact_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Path"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "evidence_type": {
            "enum": [
              "source_text",
              "source_path",
              "adapter_trace",
              "candidate_value",
              "mapping_evidence",
              "review_note",
              "knowledge_rule",
              "manifest_hash",
              "contract_check"
            ],
            "title": "Evidence Type",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "evidence_id",
          "evidence_type"
        ],
        "title": "LineageEvidence",
        "type": "object"
      },
      "LineageGraph": {
        "additionalProperties": false,
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/LineageEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "graph_id": {
            "title": "Graph Id",
            "type": "string"
          },
          "lineage_version": {
            "default": "1.0",
            "title": "Lineage Version",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/LineageNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "package_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Id"
          },
          "schema_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Id"
          },
          "summary": {
            "title": "Summary",
            "type": "object"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "graph_id",
          "doc_id",
          "generated_at",
          "nodes",
          "edges",
          "evidence"
        ],
        "title": "LineageGraph",
        "type": "object"
      },
...
      "TaskCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "options": {
            "title": "Options",
            "type": "object"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          },
          "template_version": {
            "default": "1.0.0",
            "title": "Template Version",
            "type": "string"
          }
        },
        "required": [
          "doc_id",
          "schema_id",
          "template_id"
        ],
        "title": "TaskCreateRequest",
        "type": "object"
      },
      "TaskCreateResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskCreateResponse",
        "type": "object"
      },
      "TaskDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "input_hash": {
            "title": "Input Hash",
            "type": "string"
          },
          "options": {
            "title": "Options",
            "type": "object"
          },
          "package_zip_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Zip Path"
          },
          "report_paths": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Report Paths",
            "type": "object"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          },
          "template_version": {
            "title": "Template Version",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status",
          "doc_id",
          "schema_id",
          "schema_version",
          "template_id",
          "template_version",
          "input_hash",
          "options"
        ],
        "title": "TaskDetailResponse",
        "type": "object"
      },
...
      "TaskListItem": {
        "additionalProperties": false,
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "doc_id",
          "schema_id",
          "template_id",
          "status"
        ],
        "title": "TaskListItem",
        "type": "object"
      },
      "TaskListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SchemaPack Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/v1/tasks/{task_id}/lineage": {
      "get": {
        "operationId": "get_task_lineage_api_v1_tasks__task_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageGraph"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
...
  }
}